Move CartAdapter to Application singleton
