+++ app/src/main/java/com/github/geekarist/henripotier/PotierApplication.java
@@ -16,6 +16,7 @@ public class PotierApplication extends Application {
 
     private static PotierApplication instance;
     private HenriPotier mHenriPotier;
+    private CartDatabaseHelper mDbHelper;
 
     public PotierApplication() {
         super();
@@ -40,10 +41,15 @@ public class PotierApplication extends Application {
                 .build();
 
         this.mHenriPotier = restAdapter.create(HenriPotier.class);
+        this.mDbHelper = new CartDatabaseHelper(this, null, null, 1);
     }
 
     public HenriPotier getHenriPotier() {
         return mHenriPotier;
+    }
+
+    public CartDatabaseHelper getDbHelper() {
+        return mDbHelper;
     }
 
     public interface HenriPotier {
